requirements.txt
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
jsonschema==4.19.1
python-dateutil==2.8.2
aiohttp==3.9.1
